Updates per GitHub discussion and file notes
Split up the csmim.obj.aircraft.aircraft-state.1.yaml resources listed into
multiple types per the notes in the file, based also on the proposed paths
for each of the associated resources.
Also updated audio-signals resources from two digital discretes to a single
enumeration as a proposal. This is submitted as a discusison item, and am
happy to talk it out and/or revert it to multiple discrete (bool)  resources.
Also updated phase of flight resource in csmim.obj.aircraft.flight.1. Two main
changes: first updated the enumeration to a different ARINC standard to
align with both a definition that is already in an approved standard and
that standard also doesn't include a 0-value which aligns with the CSMIM
best practices already published, and second updated descriptions based
on the published ARINC documents for state change criteria.